memo & bbox search
